feat: updated redirects and router for new page name
- Updated page filenames to new one
- Added permanent redirects
- Updated sidebar and routing
