+++ packages/app/src/domain/layout/logic/types.ts
@@ -19,7 +19,7 @@ export type VrItemKeys =
   | 'disabled_care'
   | 'elderly_at_home'
   | 'hospital_admissions'
-  | 'measures'
+  | 'current_advices'
   | 'mortality'
   | 'nursing_home_care'
   | 'positive_tests'
@@ -38,7 +38,7 @@ export type NlItemKeys =
   | 'hospital_admissions'
   | 'intensive_care_admissions'
   | 'infectious_people'
-  | 'measures'
+  | 'current_advices'
   | 'mortality'
   | 'nursing_home_care'
   | 'positive_tests'
