+++ packages/app/src/next-config/redirects/redirects.js
@@ -43,6 +43,16 @@ async function redirects() {
       permanent: false,
     },
     {
+      source: '/landelijk/maatregelen',
+      destination: '/landelijk/geldende-adviezen',
+      permanent: true,
+    },
+    {
+      source: '/veiligheidsregio/:code/maatregelen',
+      destination: '/veiligheidsregio/:code/geldende-adviezen',
+      permanent: true,
+    },
+    {
       source: '/regio',
       destination: '/veiligheidsregio',
       permanent: false,
@@ -63,15 +73,17 @@ async function redirects() {
       destination: '/veiligheidsregio/:code',
       permanent: true,
     },
-    
+
     // Redirects for former topical pages GM/VR -> dedicated GM/VR page
     {
-      source: '/actueel/gemeente/:gm(gm|GM|gM|Gm):nr(\\d{4}):slash(/{0,1}):page*',
+      source:
+        '/actueel/gemeente/:gm(gm|GM|gM|Gm):nr(\\d{4}):slash(/{0,1}):page*',
       destination: '/gemeente/GM:nr',
       permanent: false,
     },
     {
-      source: '/actueel/veiligheidsregio/:vr(vr|VR|vR|Vr):nr(\\d{2}):slash(/{0,1}):page*',
+      source:
+        '/actueel/veiligheidsregio/:vr(vr|VR|vR|Vr):nr(\\d{2}):slash(/{0,1}):page*',
       destination: '/veiligheidsregio/VR:nr',
       permanent: false,
     },
