+++ packages/app/src/pages/landelijk/geldende-adviezen.tsx
@@ -1,13 +1,10 @@
-import { useRouter } from 'next/router';
-import { AnchorTile } from '~/components/anchor-tile';
-import { Box } from '~/components/base';
+import { Box } from '~/components/base/box';
 import { RichContent } from '~/components/cms/rich-content';
 import { TileList } from '~/components/tile-list';
 import { Heading } from '~/components/typography';
 import { Layout } from '~/domain/layout/layout';
-import { VrLayout } from '~/domain/layout/vr-layout';
+import { NlLayout } from '~/domain/layout/nl-layout';
 import { LockdownTable } from '~/domain/restrictions/lockdown-table';
-import { useIntl } from '~/intl';
 import { Languages, SiteText } from '~/locale';
 import {
   createGetStaticProps,
@@ -16,34 +13,29 @@ import {
 import {
   createGetContent,
   getLastGeneratedDate,
-  selectVrData,
   getLokalizeTexts,
 } from '~/static-props/get-data';
 import { LockdownData, RoadmapData } from '~/types/cms';
-import { replaceVariablesInText } from '~/utils/replace-variables-in-text';
 import { useDynamicLokalizeTexts } from '~/utils/cms/use-dynamic-lokalize-texts';
 
-const selectLokalizeTexts = (siteText: SiteText) => ({
-  textVr: siteText.pages.measures_page.vr,
-});
-
-type LokalizeTexts = ReturnType<typeof selectLokalizeTexts>;
-
-export { getStaticPaths } from '~/static-paths/vr';
-
-type MaatregelenData = {
+type GeldendeAdviezenData = {
   lockdown: LockdownData;
   roadmap?: RoadmapData;
 };
+
+const selectLokalizeTexts = (siteText: SiteText) => ({
+  metadataTexts: siteText.pages.topical_page.nl.nationaal_metadata,
+  textNl: siteText.pages.measures_page.nl,
+});
+
+type LokalizeTexts = ReturnType<typeof selectLokalizeTexts>;
 
 export const getStaticProps = createGetStaticProps(
   ({ locale }: { locale: keyof Languages }) =>
     getLokalizeTexts(selectLokalizeTexts, locale),
   getLastGeneratedDate,
-  selectVrData(),
-  createGetContent<MaatregelenData>((context) => {
+  createGetContent<GeldendeAdviezenData>((context) => {
     const { locale } = context;
-
     return `
     {
       'lockdown': *[_type == 'lockdown']{
@@ -68,69 +60,35 @@ export const getStaticProps = createGetStaticProps(
   })
 );
 
-const RegionalRestrictions = (props: StaticProps<typeof getStaticProps>) => {
-  const { pageText, content, vrName, lastGenerated } = props;
-
-  const { commonTexts } = useIntl();
-  const { textVr } = useDynamicLokalizeTexts<LokalizeTexts>(
+const NationalRestrictions = (props: StaticProps<typeof getStaticProps>) => {
+  const { pageText, content, lastGenerated } = props;
+  const { metadataTexts, textNl } = useDynamicLokalizeTexts<LokalizeTexts>(
     pageText,
     selectLokalizeTexts
   );
-  type VRCode = keyof typeof textVr.urls;
 
   const { lockdown } = content;
 
-  const router = useRouter();
-  const code = router.query.code as unknown as VRCode;
-
-  const regioUrl = textVr.urls[code];
-
-  const metadata = {
-    ...commonTexts.veiligheidsregio_index.metadata,
-    title: replaceVariablesInText(textVr.metadata.title, {
-      safetyRegionName: vrName,
-    }),
-    description: replaceVariablesInText(textVr.metadata.title, {
-      safetyRegionName: vrName,
-    }),
-  };
-
   return (
-    <Layout {...metadata} lastGenerated={lastGenerated}>
-      <VrLayout vrName={vrName}>
+    <Layout {...metadataTexts} lastGenerated={lastGenerated}>
+      <NlLayout>
         <TileList>
           <Box as="header" spacing={4}>
-            <Heading level={1}>
-              {replaceVariablesInText(textVr.titel, {
-                safetyRegionName: vrName,
-              })}
-            </Heading>
+            <Heading level={1}>{textNl.titel}</Heading>
             {lockdown.message.description ? (
               <Box maxWidth="maxWidthText">
                 <RichContent blocks={lockdown.message.description} />
               </Box>
             ) : null}
           </Box>
-
           <Box as="article" spacing={3}>
             <Heading level={3}>{lockdown.title}</Heading>
             <LockdownTable data={lockdown} level={1} />
           </Box>
-
-          <AnchorTile
-            external
-            title={textVr.titel_aanvullendemaatregelen}
-            href={regioUrl}
-            label={replaceVariablesInText(textVr.linktext_regionpage, {
-              safetyRegionName: vrName,
-            })}
-          >
-            {textVr.toelichting_aanvullendemaatregelen}
-          </AnchorTile>
         </TileList>
-      </VrLayout>
+      </NlLayout>
     </Layout>
   );
 };
 
-export default RegionalRestrictions;
+export default NationalRestrictions;
