+++ packages/app/src/pages/veiligheidsregio/[code]/geldende-adviezen.tsx
@@ -1,10 +1,13 @@
-import { Box } from '~/components/base/box';
+import { useRouter } from 'next/router';
+import { AnchorTile } from '~/components/anchor-tile';
+import { Box } from '~/components/base';
 import { RichContent } from '~/components/cms/rich-content';
 import { TileList } from '~/components/tile-list';
 import { Heading } from '~/components/typography';
 import { Layout } from '~/domain/layout/layout';
-import { NlLayout } from '~/domain/layout/nl-layout';
+import { VrLayout } from '~/domain/layout/vr-layout';
 import { LockdownTable } from '~/domain/restrictions/lockdown-table';
+import { useIntl } from '~/intl';
 import { Languages, SiteText } from '~/locale';
 import {
   createGetStaticProps,
@@ -13,29 +16,34 @@ import {
 import {
   createGetContent,
   getLastGeneratedDate,
+  selectVrData,
   getLokalizeTexts,
 } from '~/static-props/get-data';
 import { LockdownData, RoadmapData } from '~/types/cms';
+import { replaceVariablesInText } from '~/utils/replace-variables-in-text';
 import { useDynamicLokalizeTexts } from '~/utils/cms/use-dynamic-lokalize-texts';
 
-type MaatregelenData = {
-  lockdown: LockdownData;
-  roadmap?: RoadmapData;
-};
-
 const selectLokalizeTexts = (siteText: SiteText) => ({
-  metadataTexts: siteText.pages.topical_page.nl.nationaal_metadata,
-  textNl: siteText.pages.measures_page.nl,
+  textVr: siteText.pages.measures_page.vr,
 });
 
 type LokalizeTexts = ReturnType<typeof selectLokalizeTexts>;
+
+export { getStaticPaths } from '~/static-paths/vr';
+
+type GeldendeAdviezenData = {
+  lockdown: LockdownData;
+  roadmap?: RoadmapData;
+};
 
 export const getStaticProps = createGetStaticProps(
   ({ locale }: { locale: keyof Languages }) =>
     getLokalizeTexts(selectLokalizeTexts, locale),
   getLastGeneratedDate,
-  createGetContent<MaatregelenData>((context) => {
+  selectVrData(),
+  createGetContent<GeldendeAdviezenData>((context) => {
     const { locale } = context;
+
     return `
     {
       'lockdown': *[_type == 'lockdown']{
@@ -60,35 +68,69 @@ export const getStaticProps = createGetStaticProps(
   })
 );
 
-const NationalRestrictions = (props: StaticProps<typeof getStaticProps>) => {
-  const { pageText, content, lastGenerated } = props;
-  const { metadataTexts, textNl } = useDynamicLokalizeTexts<LokalizeTexts>(
+const RegionalRestrictions = (props: StaticProps<typeof getStaticProps>) => {
+  const { pageText, content, vrName, lastGenerated } = props;
+
+  const { commonTexts } = useIntl();
+  const { textVr } = useDynamicLokalizeTexts<LokalizeTexts>(
     pageText,
     selectLokalizeTexts
   );
+  type VRCode = keyof typeof textVr.urls;
 
   const { lockdown } = content;
 
+  const router = useRouter();
+  const code = router.query.code as unknown as VRCode;
+
+  const regioUrl = textVr.urls[code];
+
+  const metadata = {
+    ...commonTexts.veiligheidsregio_index.metadata,
+    title: replaceVariablesInText(textVr.metadata.title, {
+      safetyRegionName: vrName,
+    }),
+    description: replaceVariablesInText(textVr.metadata.title, {
+      safetyRegionName: vrName,
+    }),
+  };
+
   return (
-    <Layout {...metadataTexts} lastGenerated={lastGenerated}>
-      <NlLayout>
+    <Layout {...metadata} lastGenerated={lastGenerated}>
+      <VrLayout vrName={vrName}>
         <TileList>
           <Box as="header" spacing={4}>
-            <Heading level={1}>{textNl.titel}</Heading>
+            <Heading level={1}>
+              {replaceVariablesInText(textVr.titel, {
+                safetyRegionName: vrName,
+              })}
+            </Heading>
             {lockdown.message.description ? (
               <Box maxWidth="maxWidthText">
                 <RichContent blocks={lockdown.message.description} />
               </Box>
             ) : null}
           </Box>
+
           <Box as="article" spacing={3}>
             <Heading level={3}>{lockdown.title}</Heading>
             <LockdownTable data={lockdown} level={1} />
           </Box>
+
+          <AnchorTile
+            external
+            title={textVr.titel_aanvullendemaatregelen}
+            href={regioUrl}
+            label={replaceVariablesInText(textVr.linktext_regionpage, {
+              safetyRegionName: vrName,
+            })}
+          >
+            {textVr.toelichting_aanvullendemaatregelen}
+          </AnchorTile>
         </TileList>
-      </NlLayout>
+      </VrLayout>
     </Layout>
   );
 };
 
-export default NationalRestrictions;
+export default RegionalRestrictions;
