+++ packages/cms/src/lokalize/key-mutations.csv
@@ -4,3 +4,4 @@ timestamp,action,key,document_id,move_to
 2022-08-23T09:44:54.674Z,add,common.sidebar.categories.consequences_for_healthcare.title,FV3RVBTRePmu89JK5VA4ti,__
 2022-08-23T09:44:55.689Z,add,common.sidebar.categories.actions_to_take.title,FV3RVBTRePmu89JK5VA592,__
 2022-08-23T11:59:52.631Z,move,common.nav.links.maatregelen,vLW56y4Lep4YEw1swAHFjH,common.nav.links.samenvatting
+2022-08-25T10:17:49.570Z,move,common.sidebar.metrics.measures.title,wzQp83DAUqyqV3ewG9xX0f,common.sidebar.metrics.current_advices.title
